Replaced existing import-catalog-descriptions with updated version; moved helper classes/functions into a new utilities file
